Move resetGrid() within shrink or expand checks
resetGrid() would iterate over every box in the grid to reset their
style even if they were just added. If resetGrid() is only run afer
shrinking but before expanding, it will only reset the boxes that could
need resetting, not the newly added boxes or the ones that are going to
be removed anyways.